+++ src/interface-adapters/controllers/messages/delete-message.controller.ts
@@ -2,6 +2,7 @@ import { startSpan } from "@sentry/nextjs";
 import { UnauthenticatedError } from "@src/entities/errors/auth";
 import type { ResponseProps } from "types/global";
 import { deleteMessageUseCase } from "@src/application/use-cases/messages/delete-message.use-case";
+import { getInjection } from "@di/container";
 
 function presenter(message: ResponseProps) {
   return startSpan({ name: "deleteMessage Presenter", op: "serialize" }, () => {
@@ -20,6 +21,13 @@ export async function deleteMessageController(
     async () => {
       if (!sessionId) {
         throw new UnauthenticatedError("Must be logged in to delete a message");
+      }
+
+      const authenticationService = getInjection("IAuthService");
+      const user = await authenticationService.checkIfUserExists(sessionId);
+
+      if (!user!.isAdmin) {
+        throw new UnauthenticatedError("Must be an admin to delete a message");
       }
 
       const message = await deleteMessageUseCase(id);
